+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes
 
+## 2021 Jun 29
+* Basic Receipt Storage with gossip method
+
 ## 2021 Jun 28
 
-* Expose method is_member from crate exec-membership as runtime api method is is_executor 
+* Expose method is_member from crate exec-membership as runtime api method is is_executor